requirements.txt
gunicorn
django
mongoengine
pymongo
requests
djangorestframework
django-rest-framework-mongoengine